# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SocialFlow AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{workspaceId}/social-accounts:
    get:
      operationId: listSocialAccounts
      tags: [social-accounts]
      summary: List connected social accounts
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Social accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialAccount"
    post:
      operationId: connectSocialAccount
      tags: [social-accounts]
      summary: Connect a social account
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialAccountInput"
      responses:
        "201":
          description: Connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccount"
  /workspaces/{workspaceId}/social-accounts/{id}:
    get:
      operationId: getSocialAccount
      tags: [social-accounts]
      summary: Get social account
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Social account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccount"
    delete:
      operationId: disconnectSocialAccount
      tags: [social-accounts]
      summary: Disconnect social account
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Disconnected

  # ─── POSTS ──────────────────────────────────────────────────────────────────
  /workspaces/{workspaceId}/posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List posts
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostList"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create post
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /workspaces/{workspaceId}/posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get post
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update post
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete post
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /workspaces/{workspaceId}/posts/{id}/publish:
    post:
      operationId: publishPost
      tags: [posts]
      summary: Publish post immediately
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /workspaces/{workspaceId}/posts/{id}/duplicate:
    post:
      operationId: duplicatePost
      tags: [posts]
      summary: Duplicate post
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /workspaces/{workspaceId}/calendar:
    get:
      operationId: getCalendarPosts
      tags: [posts]
      summary: Get posts grouped by date for calendar view
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Calendar posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarDay"

  # ─── MEDIA ──────────────────────────────────────────────────────────────────
...
  /workspaces/{workspaceId}/team:
    get:
      operationId: listTeamMembers
      tags: [team]
      summary: List team members
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: inviteTeamMember
      tags: [team]
      summary: Invite team member
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInviteInput"
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /workspaces/{workspaceId}/team/{memberId}:
    patch:
      operationId: updateTeamMember
      tags: [team]
      summary: Update team member role
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: removeTeamMember
      tags: [team]
      summary: Remove team member
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ─── NOTIFICATIONS ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    SocialAccount:
      type: object
      required: [id, workspaceId, platform, username, status, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        platform:
          type: string
          enum: [facebook, instagram, linkedin, pinterest, x, youtube, threads, tiktok, google_business, wordpress, medium]
        username:
          type: string
        displayName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, expired, disconnected]
        followersCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    SocialAccountInput:
      type: object
      required: [platform, username, accessToken]
      properties:
        platform:
          type: string
        username:
          type: string
        displayName:
          type: string
        profileImageUrl:
          type: string
        accessToken:
          type: string
        refreshToken:
          type: string
        platformUserId:
          type: string
        followersCount:
          type: integer

    # ─── POST SCHEMAS ──────────────────────────────────────────────────────────
    Post:
      type: object
      required: [id, workspaceId, content, status, platforms, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        content:
          type: string
        status:
          type: string
          enum: [draft, scheduled, published, failed, cancelled]
        platforms:
          type: array
          items:
            type: string
        mediaUrls:
          type: array
          items:
            type: string
        hashtags:
          type: array
          items:
            type: string
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        publishedAt:
          type: ["string", "null"]
          format: date-time
        tone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        likesCount:
          type: ["integer", "null"]
        sharesCount:
          type: ["integer", "null"]
        commentsCount:
          type: ["integer", "null"]
        impressionsCount:
          type: ["integer", "null"]
    PostList:
      type: object
      required: [posts, total, page, limit]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PostInput:
      type: object
      required: [content, platforms]
      properties:
        content:
          type: string
          minLength: 1
        platforms:
          type: array
          items:
            type: string
        mediaUrls:
          type: array
          items:
            type: string
        hashtags:
          type: array
          items:
            type: string
        scheduledAt:
          type: string
          format: date-time
        tone:
          type: string
          enum: [professional, funny, marketing, corporate, friendly]
        status:
          type: string
          enum: [draft, scheduled]
    PostUpdate:
      type: object
      properties:
        content:
          type: string
        platforms:
          type: array
          items:
            type: string
        mediaUrls:
          type: array
          items:
            type: string
        hashtags:
          type: array
          items:
            type: string
        scheduledAt:
          type: string
          format: date-time
        tone:
          type: string
        status:
          type: string
          enum: [draft, scheduled, cancelled]
    CalendarDay:
      type: object
      required: [date, posts]
      properties:
        date:
          type: string
          format: date
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"

    # ─── MEDIA SCHEMAS ─────────────────────────────────────────────────────────
...
    TeamMember:
      type: object
      required: [id, workspaceId, userId, role, status, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        userId:
          type: ["integer", "null"]
        inviteEmail:
          type: string
        role:
          type: string
          enum: [admin, editor, viewer, client]
        status:
          type: string
          enum: [active, pending, removed]
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    TeamInviteInput:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
        role:
          type: string
          enum: [admin, editor, viewer, client]
    TeamMemberUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, editor, viewer, client]

    # ─── NOTIFICATION SCHEMAS ──────────────────────────────────────────────────
...